DEV: Extract textarea text manipulation to mixin
